Exploratory data analysis with python
This files contain all codes and plots used in exploratory data analysis project. Each code written is a contributed to by Alex the Analyst.
This project is to help understand the data before the eventual data analysis process especially the data cleaning/scrubbing phase which is necessary for good and accurate analysis. 
What is the use of a data analysis if the data is not fit?