Exploratory Data Analysis using Python
This file contains all code used in this project with the aid of Alex the Analyst as a major contributor to this work.
Data Exploratory is to discover and be familiar with the data you are about to work on. It's to identify patterns, outliers and prepare for the data cleaning/scrubbing phase in analysis project.
what's the use of a data or it's analysis if the data does not have good data quality like data accuracy?